Dynamic nav bar, add post.php support for logged in users

+++ database/postFunctions.php
@@ -22,6 +22,30 @@ function convertToLocal($datetime)
 	return $date->format('m/d/Y \a\t g:ia');
 }
 
+function get_post_author($post_id)
+{
+	$sql = "SELECT author_id FROM posts WHERE post_id=:post_id";
+	$params = [":post_id" => $post_id];
+
+	try {
+		return getDataFromSQL($sql, $params)[0]["author_id"];
+	} catch (Exception $e) {
+		header("HTTP/1.1 500 Fatal Error");
+	}
+}
+
+function get_comment_author($comment_id)
+{
+	$sql = "SELECT author_id FROM comments WHERE comment_id=:comment_id";
+	$params = [":comment_id" => $comment_id];
+
+	try {
+		return getDataFromSQL($sql, $params)[0]["author_id"];
+	} catch (Exception $e) {
+		header("HTTP/1.1 500 Fatal Error");
+	}
+}
+
 function getPost($post_id, $includeExtraData = false)
 {
 	$sql = "SELECT post_id,creationDate,title,content,username,author_id FROM post_with_username WHERE inactive=0 AND post_id=:id";
@@ -40,11 +64,13 @@ function getPost($post_id, $includeExtraData = false)
 
 function createPost($title, $content, $zip)
 {
+	if (!isset($_SESSION["user"]))
+		return header("HTTP/1.1 401 Unauthorized");
+
+	$author_id = $_SESSION["user"];
+
 	$sql = "INSERT INTO posts (author_id, title, content, zip)
 	VALUES (:author_id, :title, :content, :zip)";
-
-	// Use 2 as the default id until we have session data
-	$author_id = 2;
 
 	$params = [
 		":author_id" => $author_id,
@@ -62,12 +88,20 @@ function createPost($title, $content, $zip)
 
 function editPost($post_id, $title, $content, $zip)
 {
+	$actual_author = get_post_author($post_id);
+
+	if (!isset($_SESSION["user"]) || $_SESSION["user"] != $actual_author)
+		return header("HTTP/1.1 401 Unauthorized");
+
+	$author_id = $_SESSION["user"];
+
 	$sql = "UPDATE posts 
 	SET title=:title, content=:content, zip=:zip
-	WHERE post_id=:post_id";
+	WHERE post_id=:post_id AND author_id=:author_id";
 
 	$params = [
 		":post_id" => $post_id,
+		":author_id" => $author_id,
 		":title" => $title,
 		":content" => $content,
 		":zip" => $zip
@@ -82,11 +116,21 @@ function editPost($post_id, $title, $content, $zip)
 
 function deletePost($post_id)
 {
+	$actual_author = get_post_author($post_id);
+
+	if (!isset($_SESSION["user"]) || $_SESSION["user"] != $actual_author)
+		return header("HTTP/1.1 401 Unauthorized");
+
+	$author_id = $_SESSION["user"];
+
 	$sql = "UPDATE posts 
 	SET inactive=1
-	WHERE post_id=:post_id";
+	WHERE post_id=:post_id AND author_id=:author_id";
 
-	$params = [":post_id" => $post_id];
+	$params = [
+		":post_id" => $post_id,
+		":author_id" => $author_id,
+	];
 
 	try {
 		postDataFromSQL($sql, $params);
@@ -99,19 +143,22 @@ function getPosts($type): array
 {
 	$sql = "SELECT post_id,creationDate,title,username,author_id,zip FROM post_with_username WHERE inactive=0";
 
-	// Use 2 as the default id until we have session data
-	$user_id = 2;
-
-	$params = [
-		":user_id" => $user_id
-	];
+	$params = null;
 
 	if ($type == "your") {
 		$sql = $sql . " AND author_id=:user_id";
 	} else if ($type == "saved")
 		$sql = $sql . " AND post_id IN (SELECT post_id FROM saves WHERE user_id=:user_id)";
-	else
-		$params = null;
+	else $type = null;
+
+	if ($type) {
+		if (isset($_SESSION["user"]))
+			$params = [
+				":user_id" => $_SESSION["user"]
+			];
+		else
+			return header("HTTP/1.1 401 Unauthorized");
+	}
 
 	$posts = getDataFromSQL($sql, $params);
 
@@ -155,11 +202,13 @@ function getRepliesToComment($comment_id): array
 
 function addComment($post_id, $content, $parent_id = null)
 {
+	if (!isset($_SESSION["user"]))
+		return header("HTTP/1.1 401 Unauthorized");
+
+	$author_id = $_SESSION["user"];
+
 	$sql = "INSERT INTO comments (post_id, author_id, content, parent_id)
           VALUES (:post_id, :author_id, :content, :parent_id)";
-
-	// Use 2 as the default id until we have session data
-	$author_id = 2;
 
 	$params =
 		[
@@ -178,11 +227,13 @@ function addComment($post_id, $content, $parent_id = null)
 
 function applicationExists($post_id)
 {
+	if (!isset($_SESSION["user"]))
+		return header("HTTP/1.1 401 Unauthorized");
+
+	$user_id = $_SESSION["user"];
+
 	$sql = "SELECT COUNT(post_id) as \"count\" FROM applications
           WHERE post_id=:post_id AND user_id=:user_id";
-
-	// Use 2 as the default id until we have session data
-	$user_id = 2;
 
 	$params =
 		[
@@ -201,15 +252,19 @@ function applicationExists($post_id)
 
 function apply($post_id)
 {
+	$actual_author = get_post_author($post_id);
+
+	if (!isset($_SESSION["user"]) || $_SESSION["user"] == $actual_author)
+		return header("HTTP/1.1 401 Unauthorized");
+
+	$user_id = $_SESSION["user"];
+
 	if (applicationExists($post_id)) {
 		return header("HTTP/1.1 403 Already Exists");
 	}
 
 	$sql = "INSERT INTO applications (post_id, user_id)
           VALUES (:post_id, :user_id)";
-
-	// Use 2 as the default id until we have session data
-	$user_id = 2;
 
 	$params =
 		[
@@ -226,11 +281,13 @@ function apply($post_id)
 
 function saveExists($post_id)
 {
+	if (!isset($_SESSION["user"]))
+		return header("HTTP/1.1 401 Unauthorized");
+
+	$user_id = $_SESSION["user"];
+
 	$sql = "SELECT COUNT(post_id) as \"count\" FROM saves
           WHERE post_id=:post_id AND user_id=:user_id";
-
-	// Use 2 as the default id until we have session data
-	$user_id = 2;
 
 	$params =
 		[
@@ -249,15 +306,19 @@ function saveExists($post_id)
 
 function save($post_id)
 {
+	$actual_author = get_post_author($post_id);
+
+	if (!isset($_SESSION["user"]) || $_SESSION["user"] == $actual_author)
+		return header("HTTP/1.1 401 Unauthorized");
+
+	$user_id = $_SESSION["user"];
+
 	if (saveExists($post_id)) {
 		return header("HTTP/1.1 403 Already Exists");
 	}
 
 	$sql = "INSERT INTO saves (post_id, user_id)
           VALUES (:post_id, :user_id)";
-
-	// Use 2 as the default id until we have session data
-	$user_id = 2;
 
 	$params =
 		[
@@ -274,15 +335,19 @@ function save($post_id)
 
 function unsave($post_id)
 {
+	$actual_author = get_post_author($post_id);
+
+	if (!isset($_SESSION["user"]) || $_SESSION["user"] == $actual_author)
+		return header("HTTP/1.1 401 Unauthorized");
+
+	$user_id = $_SESSION["user"];
+
 	if (!saveExists($post_id)) {
 		return header("HTTP/1.1 403 Does Not Exist");
 	}
 
 	$sql = "DELETE FROM saves
           WHERE post_id=:post_id AND user_id=:user_id";
-
-	// Use 2 as the default id until we have session data
-	$user_id = 2;
 
 	$params =
 		[
@@ -299,11 +364,13 @@ function unsave($post_id)
 
 function reportExists($id, $type)
 {
+	if (!isset($_SESSION["user"]))
+		return header("HTTP/1.1 401 Unauthorized");
+
+	$user_id = $_SESSION["user"];
+
 	$sql = "SELECT COUNT(id) as \"count\" FROM reports
           WHERE id=:id AND type=:type AND reporter_id=:user_id";
-
-	// Use 2 as the default id until we have session data
-	$user_id = 2;
 
 	$params =
 		[
@@ -323,15 +390,20 @@ function reportExists($id, $type)
 
 function report($id, $type)
 {
+	$actual_author = $type == 1
+		? get_post_author($id) : get_comment_author($id);
+
+	if (!isset($_SESSION["user"]) || $_SESSION["user"] == $actual_author)
+		return header("HTTP/1.1 401 Unauthorized");
+
+	$user_id = $_SESSION["user"];
+
 	if (reportExists($id, $type)) {
 		return header("HTTP/1.1 403 Already Exists");
 	}
 
 	$sql = "INSERT INTO reports (reporter_id, id, type)
           VALUES (:reporter_id, :id, :type)";
-
-	// Use 2 as the default id until we have session data
-	$user_id = 2;
 
 	$params =
 		[
